Move httpStatusCodes, httpMethods and Url within top-level proxy.http package
