+++ examples/https_connect_tunnel.py
@@ -14,7 +14,8 @@ from typing import Any, Optional
 
 from proxy import Proxy
 from proxy.common.utils import build_http_response
-from proxy.http.parser import httpParserStates, httpStatusCodes
+from proxy.http import httpStatusCodes
+from proxy.http.parser import httpParserStates
 from proxy.core.base import BaseTcpTunnelHandler
