+++ proxy/core/connection/client.py
@@ -8,8 +8,9 @@
     :copyright: (c) 2013-present by Abhinav Singh and contributors.
     :license: BSD, see LICENSE for more details.
 """
-import socket
 import ssl
+import socket
+
 from typing import Union, Tuple, Optional
 
 from .connection import TcpConnection, TcpConnectionUninitializedException
