+++ proxy/core/connection/connection.py
@@ -8,8 +8,8 @@
     :copyright: (c) 2013-present by Abhinav Singh and contributors.
     :license: BSD, see LICENSE for more details.
 """
-import socket
 import ssl
+import socket
 import logging
 
 from abc import ABC, abstractmethod
