+++ proxy/core/connection/pool.py
@@ -13,6 +13,7 @@ import logging
 from typing import Set, Dict, Tuple
 
 from ...common.flag import flags
+
 from .server import TcpServerConnection
 
 logger = logging.getLogger(__name__)
