+++ proxy/http/__init__.py
@@ -10,8 +10,14 @@
 """
 from .handler import HttpProtocolHandler
 from .plugin import HttpProtocolHandlerPlugin
+from .codes import httpStatusCodes
+from .methods import httpMethods
+from .url import Url
 
 __all__ = [
     'HttpProtocolHandler',
     'HttpProtocolHandlerPlugin',
+    'httpStatusCodes',
+    'httpMethods',
+    'Url',
 ]
