+++ proxy/plugin/filter_by_upstream.py
@@ -12,9 +12,11 @@ from typing import Optional
 
 from ..common.utils import text_
 from ..common.flag import flags
-from ..http.exception import HttpRequestRejected
-from ..http.parser import HttpParser, httpStatusCodes
+
+from ..http import httpStatusCodes
+from ..http.parser import HttpParser
 from ..http.proxy import HttpProxyBasePlugin
+from ..http.exception import HttpRequestRejected
 
 
 flags.add_argument(
