+++ proxy/plugin/filter_by_url_regex.py
@@ -14,10 +14,12 @@ import logging
 from typing import Optional, List, Dict, Any
 
 from ..common.flag import flags
-from ..http.exception import HttpRequestRejected
-from ..http.parser import HttpParser, httpStatusCodes
-from ..http.proxy import HttpProxyBasePlugin
 from ..common.utils import text_
+
+from ..http import httpStatusCodes
+from ..http.parser import HttpParser
+from ..http.proxy import HttpProxyBasePlugin
+from ..http.exception import HttpRequestRejected
 
 import re
