+++ proxy/plugin/man_in_the_middle.py
@@ -8,10 +8,8 @@
     :copyright: (c) 2013-present by Abhinav Singh and contributors.
     :license: BSD, see LICENSE for more details.
 """
-from typing import Optional
-
 from ..common.utils import build_http_response
-from ..http.parser import httpStatusCodes
+from ..http import httpStatusCodes
 from ..http.proxy import HttpProxyBasePlugin
