+++ proxy/plugin/mock_rest_api.py
@@ -12,7 +12,9 @@ import json
 from typing import Optional
 
 from ..common.utils import bytes_, build_http_response, text_
-from ..http.parser import HttpParser, httpStatusCodes
+
+from ..http import httpStatusCodes
+from ..http.parser import HttpParser
 from ..http.proxy import HttpProxyBasePlugin
