+++ proxy/plugin/modify_post_data.py
@@ -11,7 +11,9 @@
 from typing import Optional
 
 from ..common.utils import bytes_
-from ..http.parser import HttpParser, httpMethods
+
+from ..http import httpMethods
+from ..http.parser import HttpParser
 from ..http.proxy import HttpProxyBasePlugin
