+++ proxy/plugin/shortlink.py
@@ -12,7 +12,9 @@ from typing import Optional
 
 from ..common.constants import DOT, SLASH
 from ..common.utils import build_http_response
-from ..http.parser import HttpParser, httpStatusCodes
+
+from ..http import httpStatusCodes
+from ..http.parser import HttpParser
 from ..http.proxy import HttpProxyBasePlugin
