+++ tests/http/exceptions/test_http_request_rejected.py
@@ -10,7 +10,8 @@
 """
 import unittest
 
-from proxy.http.parser import HttpParser, httpParserTypes, httpStatusCodes
+from proxy.http import httpStatusCodes
+from proxy.http.parser import HttpParser, httpParserTypes
 from proxy.http.exception import HttpRequestRejected
 from proxy.common.constants import CRLF
 from proxy.common.utils import build_http_response
