+++ tests/plugin/test_http_proxy_plugins_with_tls_interception.py
@@ -19,9 +19,9 @@ from typing import Any, cast
 from proxy.common.flag import FlagParser
 from proxy.common.utils import bytes_, build_http_request, build_http_response
 from proxy.core.connection import TcpClientConnection, TcpServerConnection
+
+from proxy.http import httpMethods, httpStatusCodes, HttpProtocolHandler
 from proxy.http.proxy import HttpProxyPlugin
-from proxy.http.parser import httpStatusCodes, httpMethods
-from proxy.http import HttpProtocolHandler
 
 from .utils import get_plugin_by_test_name
