+++ tests/testing/test_embed.py
@@ -17,8 +17,8 @@ import urllib.error
 from proxy import TestCase
 from proxy.common.constants import DEFAULT_CLIENT_RECVBUF_SIZE, PROXY_AGENT_HEADER_VALUE
 from proxy.common.utils import socket_connection, build_http_request
+from proxy.http import httpMethods
 from proxy.http.server import HttpWebServerPlugin
-from proxy.http.parser import httpMethods
 
 
 @unittest.skipIf(os.name == 'nt', 'Disabled for Windows due to weird permission issues.')
